requirements/dev.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.0.1
sqlformatter==1.0
ipython==2.0.0
ipdb==0.8
coverage==3.7.1
django-nose==1.2
django-extensions==1.3.7
Werkzeug==0.9.4
model_mommy